% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CC_Pheno_Cor.R
\docType{data}
\name{CC_Pheno_Cor}
\alias{CC_Pheno_Cor}
\title{Clinical Chemistry Phenotype Correlation Matrix}
\format{
An object of class 'matrix' with 19 rows and 19 columns, where each cell
contains the correlation value between two different phenotypes.
}
\usage{
data(CC_Pheno_Cor)
}
\description{
This data set is a correlation matrix between 19 Clinical Chemistry
phenotypes. These phenotypes are derived from the version 10.1 release of the
International Mouse Phenotyping Consortium (IMPC) Control Phenotype data.
Each cell in the matrix represents the correlation between a pair of these phenotypes.
Notably, before calculating the correlation coefficients, we performed a regression analysis
to eliminate the effect of variation between different phenotyping centers.
}
\examples{
\dontrun{
data(CC_Pheno_Cor)
str(CC_Pheno_Cor)
}

}
\references{
Warkentin et al. (2023). KOMPUTE: Imputing summary statistics of missing
phenotypes in high-throughput model organism data. bioRxiv.
\href{https://www.biorxiv.org/content/10.1101/2023.01.12.523855v1}{Link to article}
}
\keyword{datasets}
