% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kompute.R
\name{kompute}
\alias{kompute}
\title{Imputation of association Z-scores of missing phenotypes}
\usage{
kompute(zmat, pheno.cor, ridge = 0.001)
}
\arguments{
\item{zmat}{The matrix of Z-scores, row: number of phenotypes, col: number of genes}

\item{pheno.cor}{The phenotype correlation matrix}

\item{ridge}{The ridge panelty}
}
\description{
This fuction estimates association Z-scores of unmeasured phenotypes.
}
